Đơn giản hóa màn 2
Với phần khung nền mới được thêm vào có thể tách các đoạn code trùng lặp
giữa các bài học ra. Các bài học khác có thể làm tương tự màn 2.
